Cleanup older cli tests
